+++ paddle/phi/kernels/cpu/unpool_kernel.cc
@@ -33,12 +33,12 @@ void Unpool(const Context& dev_ctx,
     phi::funcs::SetConstant<Context, T> set_zero;
     set_zero(dev_ctx, out, static_cast<T>(0));
   }
-  const int64_t batch_size = static_cast<int64_t>(x.dims()[0]);
-  const int64_t input_height = static_cast<int64_t>(x.dims()[2]);
-  const int64_t input_width = static_cast<int64_t>(x.dims()[3]);
-  const int64_t output_channels = static_cast<int64_t>(out->dims()[1]);
-  const int64_t output_height = static_cast<int64_t>(out->dims()[2]);
-  const int64_t output_width = static_cast<int64_t>(out->dims()[3]);
+  const int64_t batch_size = x.dims()[0];
+  const int64_t input_height = x.dims()[2];
+  const int64_t input_width = x.dims()[3];
+  const int64_t output_channels = out->dims()[1];
+  const int64_t output_height = out->dims()[2];
+  const int64_t output_width = out->dims()[3];
   int64_t input_feasize = input_height * input_width;
   int64_t output_feasize = output_height * output_width;
   const T* input_data = x.data<T>();
@@ -99,14 +99,14 @@ void Unpool3d(const Context& dev_ctx,
     phi::funcs::SetConstant<Context, T> set_zero;
     set_zero(dev_ctx, out, static_cast<T>(0));
   }
-  const int64_t batch_size = static_cast<int64_t>(x.dims()[0]);
-  const int64_t input_depth = static_cast<int64_t>(x.dims()[2]);
-  const int64_t input_height = static_cast<int64_t>(x.dims()[3]);
-  const int64_t input_width = static_cast<int64_t>(x.dims()[4]);
-  const int64_t output_channels = static_cast<int64_t>(out->dims()[1]);
-  const int64_t output_depth = static_cast<int64_t>(out->dims()[2]);
-  const int64_t output_height = static_cast<int64_t>(out->dims()[3]);
-  const int64_t output_width = static_cast<int64_t>(out->dims()[4]);
+  const int64_t batch_size = x.dims()[0];
+  const int64_t input_depth = x.dims()[2];
+  const int64_t input_height = x.dims()[3];
+  const int64_t input_width = x.dims()[4];
+  const int64_t output_channels = out->dims()[1];
+  const int64_t output_depth = out->dims()[2];
+  const int64_t output_height = out->dims()[3];
+  const int64_t output_width = out->dims()[4];
   int64_t input_feasize = input_depth * input_height * input_width;
   int64_t output_feasize = output_depth * output_height * output_width;
   const T* input_data = x.data<T>();
